requirements.txt
openai
pydantic
pytest
python-dotenv